+++ electron/collection.css
@@ -1,97 +1,20 @@
-h4 {
-    line-height: 1.4;
-}
-
-.settings-container {
-  padding: 10px;
-  margin: 10px;
-}
-
-.settings-row {
-  margin-top: 10px;
-}
-
-.toggle {
-  min-width: 65px;
-  min-height: 40px;
-}
-
-.about-p {
-  padding: 0;
-  margin: 0;
-}
-
-a {
-  cursor: pointer;
+.pane {
+  padding: 15px;
 }
 
 #sidebar {
   max-width: 180px;
+  padding: 10px 0 0 0;
 }
 
-.link {
-  cursor: pointer;
-}
-
-.btnbtn {
-  margin-top: 12px;
-}
-
-.helper-text, .helper-block {
-  font-style: italic;
-}
-
-.helper-text {
-}
-
-.helper-block {
-  margin: 5px 0 0 0;
-  display: block;
-}
-
-span.helper-text::before {
-    content: "- ";
-}
-
-#sorting-method-select {
-    margin: 5px 0 8px 0;
-}
-
-.toggle {
-  margin-right: 5px;
-}
-
-.disclaimer {
-  font-size: 14px;
-  font-style: italic;
-}
-
-.well {
-    min-height: 20px;
-    padding: 19px;
-    margin-bottom: 20px;
-    background-color: #f5f5f5;
-    border: 1px solid #e3e3e3;
-    border-radius: 4px;
-    -webkit-box-shadow: inset 0 1px 1px rgba(0,0,0,.05);
-    box-shadow: inset 0 1px 1px rgba(0,0,0,.05);
-}
-
-.maj-img {
-  width: 140px;
-  height: 140px;
-  border-radius: 7px;
-}
-
-.maj-container {
-  display: inline-block;
-  width: 210px;
-  height: 240px;
-}
-
-.small-hr {
-  margin-top: 0.8rem;
-  margin-bottom: 0.2rem;
+.fa {
+  width: 19px;
+  height: 18px;
+  color: #737475;
+  margin-right: 2px;
+  margin-left: 1px;
+  font-size: 16px;
+  text-align: center;
 }
 
 .slider-label {
@@ -142,35 +65,15 @@ span.helper-text::before {
     cursor: pointer;
 }
 
-.authorize-button, .authorize-submit-button {
-  float: right;
-  height: 23px;
-  font-size: 12px!important;
-  margin: 0 5px;
-  padding-top: 2px!important;
-}
-
-.authorize-submit-button{
-  margin-top: 8px;
-}
-
-.account-entry {
-  padding: 5px 5px;
-  border-top: solid rgb(220,220,220) 1px;
-  border-bottom: solid rgb(220,220,220) 1px;
-}
-
-.account-entry:first-child {
-  border-top: solid rgb(220,220,220) 2px;
-}
-
-.account-entry:last-child {
-  border-bottom: solid rgb(220,220,220) 2px;
-}
-
-.access-container {
-  display: none;
-  padding-top: 10px;
+.well {
+    min-height: 20px;
+    padding: 19px;
+    margin-bottom: 20px;
+    background-color: #f5f5f5;
+    border: 1px solid #e3e3e3;
+    border-radius: 4px;
+    -webkit-box-shadow: inset 0 1px 1px rgba(0,0,0,.05);
+    box-shadow: inset 0 1px 1px rgba(0,0,0,.05);
 }
 
 .recent-cards-group-container {
@@ -241,72 +144,6 @@ span.helper-text::before {
   background-color: black;
 }
 
-/* https://codepen.io/Deadlymuffin/pen/hGiqo/ */
-
-.social-icon {
-    color: #fff;
-    background-color: #aaa;
-    padding: 3px 15px;
-    border-radius: 4px;
-    -o-transition:.5s;
-    -ms-transition:.5s;
-    -moz-transition:.5s;
-    -webkit-transition:.5s;
-}
-.social-icon:hover {
-  cursor: pointer;
-}
-.fa-facebook:hover {
-    background-color: #3d5b99;
-}
-.fa-github:hover {
-    background-color: #333;
-}
-.fa-link:hover {
-    background-color: #333;
-}
-.fa-twitter:hover {
-    background-color: #00aced;
-}
-.fa-twitch:hover {
-    background-color: #6441A4;
-    color: white;
-}
-.fa-rss:hover {
-    background-color: #eb8231;
-}
-.fa-youtube:hover {
-    background-color: #e64a41;
-}
-.fa-linkedin:hover {
-    background-color: #0073a4;
-}
-.fa-google-plus:hover {
-    background-color: #e25714;
-}
-
-.counter-toggle-span {
-    vertical-align: middle;
-    margin-left: 10px;
-}
-
-.counter-reset-span {
-    margin-top: 8px;
-    display: inline-block;
-}
-
-.reset-all-btn {
-    width: 20em;
-}
-
-.counter-adj-span {
-    float: right;
-}
-
-.counter-reset-row {
-  height: 32px;
-}
-
 .inv-label img {
     width: 32px;
     margin-right: 5px;
@@ -351,22 +188,6 @@ td.inv-label {
     margin-left: 5px;
 }
 
-.wcCommon {
-  background-color: #bebebe;
-}
-
-.wcUncommon {
-  background-color: #a4c1d1;
-}
-
-.wcRare {
-  background-color: #fbca2a;
-}
-
-.wcMythic {
-  background-color: #ff6c0d;
-}
-
 .gained {
   color: green;
 }
@@ -382,3 +203,6 @@ th.table-type {
   font-weight: bold;
 }
 
+h3 {
+  line-height: 1.3;
+}
