+++ electron/settingsRenderer.js
@@ -15,8 +15,6 @@ rendererPreload();
 
 var { databaseFiles } = require("./conf")
 
-let desktopPath = path.join(os.homedir(), 'Desktop')
-
 const tt = require('electron-tooltip')
 tt({
   position: 'top',
@@ -38,10 +36,6 @@ var buildWorker = function(func) {
   return worker
 }
 
-let inventory = remote.getGlobal('inventory')
-let inventorySpent = remote.getGlobal('inventorySpent')
-let inventoryGained = remote.getGlobal('inventoryGained')
-
 var settingsData = {
   version: remote.getGlobal("version"),
   commit: "",
@@ -53,6 +47,7 @@ var settingsData = {
   debug: remote.getGlobal('debug'),
   showErrors: remote.getGlobal('showErrors'),
   showInspector: remote.getGlobal('showInspector'),
+  sendAnonymousUsageInfo: remote.getGlobal('sendAnonymousUsageInfo'),
   incognito: remote.getGlobal('incognito'),
   useFrame: remote.getGlobal('useFrame'),
   staticMode: remote.getGlobal('staticMode'),
@@ -69,10 +64,6 @@ var settingsData = {
   dailyCounterDeckList: [],
   totalWinLossCounter: null,
   dailyTotalWinLossCounter: null,
-  lastCollection: remote.getGlobal('lastCollection'),
-  lastCollectionCount: "loading...",
-  lastCollectionSetProgress: [{name: "loading..."}],
-  lastVaultProgress: remote.getGlobal('lastVaultProgress'),
   showVaultProgress: remote.getGlobal('showVaultProgress'),
   minVaultProgress: remote.getGlobal('minVaultProgress'),
   showGameTimer: remote.getGlobal('showGameTimer'),
@@ -80,8 +71,6 @@ var settingsData = {
   hideDelay: remote.getGlobal('hideDelay'),
   invertHideMode: remote.getGlobal('invertHideMode'),
   rollupMode: remote.getGlobal('rollupMode'),
-  recentCards: remote.getGlobal('recentCards'),
-  recentCardsQuantityToShow: remote.getGlobal('recentCardsQuantityToShow'),
   minToTray: remote.getGlobal('minToTray'),
   runFromSource: remote.getGlobal('runFromSource'),
   sortMethodSelected: remote.getGlobal('sortMethod'),
@@ -105,26 +94,7 @@ var settingsData = {
   ],
   showUIButtons: remote.getGlobal('showUIButtons'),
   showHideButton: remote.getGlobal('showHideButton'),
-  showMenu: remote.getGlobal('showMenu'),
-  gold: inventory.gold,
-  goldSpent: inventorySpent.gold,
-  goldGained: inventoryGained.gold,
-  gems: inventory.gems,
-  gemsSpent: inventorySpent.gems,
-  gemsGained: inventoryGained.gems,
-  wcCommon: inventory.wcCommon,
-  wcCommonSpent: inventorySpent.wcCommon,
-  wcCommonGained: inventoryGained.wcCommon,
-  wcUncommon: inventory.wcUncommon,
-  wcUncommonSpent: inventorySpent.wcUncommon,
-  wcUncommonGained: inventoryGained.wcUncommon,
-  wcRare: inventory.wcRare,
-  wcRareSpent: inventorySpent.wcRare,
-  wcRareGained: inventoryGained.wcRare,
-  wcMythic: inventory.wcMythic,
-  wcMythicSpent: inventorySpent.wcMythic,
-  wcMythicGained: inventoryGained.wcMythic,
-  boosters: inventory.boosters
+  showMenu: remote.getGlobal('showMenu')
 }
 
 settingsData.counterDeckList = counterDecks(settingsData.winLossObj.alltime);
@@ -172,15 +142,6 @@ ipcRenderer.on('counterChanged', (e,new_wlc) => {
   settingsData.totalWinLossCounter = settingsData.winLossObj.alltime.total;
   settingsData.dailyTotalWinLossCounter = settingsData.winLossObj.daily.total;
 });
-
-ipcRenderer.on('inventoryChanged',(e,new_inventory,new_inventory_spent,new_inventory_gained) => {
-  let fields = ['gold','gems','wcCommon','wcUncommon','wcRare','wcMythic']
-  for (let field of fields){
-    settingsData[field] = new_inventory[field]
-    settingsData[field + 'Spent'] = new_inventory_spent[field]
-    settingsData[field + 'Gained'] = new_inventory_gained[field]
-  }
-})
 
 /*
  * Format decks in winLossCounter to array for display in rivets.
@@ -265,224 +226,13 @@ rivets.binders.authref = (el, val) => {
   el.href = "https://inspector.mtgatracker.com/trackerAuth?code=" + val
 }
 
-rivets.binders.recentcardsbinder = (el, cardsObtained) => {
-  var node;
-  var textNode;
-  var currentCard
-  for(var cardID in cardsObtained) {
-    currentCard = mtga.allCards.findCard(cardID)
-    if(currentCard) {
-      textNode = document.createTextNode(`${cardsObtained[cardID]}x ${currentCard.attributes.prettyName}`);
-    } else {
-      textNode = document.createTextNode(`${cardsObtained[cardID]}x card-name-not-found (${cardID})`);
-    }
-    node = document.createElement("li");
-    node.style.webkitUserSelect = "auto";
-    node.appendChild(textNode);
-    el.appendChild(node);
-  }
-  if(Object.keys(cardsObtained).length > 0) {
-    document.getElementById("no-recently-obtained-cards").style.display = "none";
-  }
-}
-
 rivets.binders.deckid = function(el, value) {
   el.setAttribute('data-deckid', value);
-}
-
-rivets.binders.mythicprogress = function(el, value) {
-  el.style.width = Math.max(0, (100 * value.mythicOwned / value.mythicTotal)) + "%"
-}
-
-rivets.binders.rareprogress = function(el, value) {
-  el.style.width = Math.max(0, (100 * value.rareOwned / value.rareTotal)) + "%"
-}
-
-rivets.binders.uncommonprogress = function(el, value) {
-  el.style.width = Math.max(0, (100 * value.uncommonOwned / value.uncommonTotal)) + "%"
-}
-
-rivets.binders.commonprogress = function(el, value) {
-  el.style.width = Math.max(0, (100 * value.commonOwned / value.commonTotal)) + "%"
-}
-
-rivets.binders.netinv = (el,value) => {
-  let $el = $(el)
-  $el.text(value)
-  if (value > 0){
-    $el.addClass('gained')
-    $el.removeClass('spent')
-  } else if (value < 0) {
-    $el.addClass('spent')
-    $el.removeClass('gained')
-  } else {
-    $el.removeClass('spent')
-    $el.removeClass('gained')
-  }
-
-}
-
-settingsData.netGold = () => {
-  return settingsData.goldGained - settingsData.goldSpent
-}
-
-settingsData.netGems = () => {
-  return settingsData.gemsGained - settingsData.gemsSpent
-}
-
-settingsData.netDraftTokens = () => {
-  return settingsData.draftTokensGained - settingsData.draftTokensSpent
-}
-
-settingsData.netWcCommon = () => {
-  return settingsData.wcCommonGained - settingsData.wcCommonSpent
-}
-
-settingsData.netWcUncommon = () => {
-  return settingsData.wcUncommonGained - settingsData.wcUncommonSpent
-}
-
-settingsData.netWcRare = () => {
-  return settingsData.wcRareGained - settingsData.wcRareSpent
-}
-
-settingsData.netWcMythic = () => {
-  return settingsData.wcMythicGained - settingsData.wcMythicSpent
-}
-
-const setPromoMap = {
-  RIX: "img/card_set_promos/rix.png",
-  M19: "img/card_set_promos/m19.png",
-  GRN: "img/card_set_promos/grn.png",
-  XLN: "img/card_set_promos/xln.png",
-  DAR: "img/card_set_promos/dar.png",
-  ANA: "img/card_set_promos/ana.png",
-}
-
-rivets.binders.setpromo = function(el, value) {
-  if (Object.keys(setPromoMap).includes(value)) {
-    el.style.display = "block"
-    el.src = setPromoMap[value]
-  } else {
-    el.style.display = "none"
-  }
-}
-
-rivets.binders.hidesetname = function(el, value) {
-  if (Object.keys(setPromoMap).includes(value)) {
-    el.style.display = "none"
-  } else {
-    el.style.display = "block"
-  }
-}
-
-function recentCardsSectionClickHandler(event) {
-  var revealed = $(event.target).siblings(".recent-cards-container").is(":hidden");
-  if(revealed) {
-    $(event.target).siblings(".recent-cards-container").slideDown("fast");
-  } else {
-    $(event.target).siblings(".recent-cards-container").slideUp("fast");
-  }
 }
 
 var firstCounterAdjButtonClicked = true;
 document.addEventListener("DOMContentLoaded", function(event) {
   rivets.bind(document.getElementById('container'), settingsData)
-
-  let collectionWorker = buildWorker(e => {
-    var allCards;
-    var cardSets = {}
-    var playerCardCounts = {}
-    onmessage = event => {
-      if (event.data.allCards) {
-        allCards = event.data.allCards.attributes.cards;
-        // Note: this block of code looks really stupid, but trust me, it's necessary.
-        // TL:DR; you can't `require(...)` inside webworkers, so we lose all of the cool mtga functionality.
-        // As if that wasn't bad enough: since mtga uses backbone BS, we have to do silly things to
-        // get to the actual objects within the allCards object.
-        // Anyways, this block of code basically redoes all the organization that mtga originally offered in the
-        // first place. :tiny_violin:
-        for (let cardID in allCards) {
-          let thisCard = allCards[cardID].attributes
-          if (!cardSets[thisCard.set]) {
-            cardSets[thisCard.set] = {
-              cards: [],
-              counts: {
-                mythicTotal: 0,
-                rareTotal: 0,
-                uncommonTotal: 0,
-                commonTotal: 0
-              }
-            }
-          }
-          let thisCardsSet = cardSets[thisCard.set]
-          thisCardsSet.cards.push(thisCard)
-          // add 4 for each unique card; you can collect 4 of each
-          if (thisCard.rarity == "Mythic Rare") {
-            thisCardsSet.counts.mythicTotal += 4;
-          } else if (thisCard.rarity == "Rare") {
-            thisCardsSet.counts.rareTotal += 4;
-          } else if (thisCard.rarity == "Uncommon") {
-            thisCardsSet.counts.uncommonTotal += 4;
-          } else if (thisCard.rarity == "Common") {
-            thisCardsSet.counts.commonTotal += 4;
-          }
-        }
-        console.log(cardSets)
-        postMessage({ready: true})
-      } else if (event.data.lastCollection) {
-        let total = 0;
-        let unique = 0;
-        let collection = event.data.lastCollection;
-        console.log(Object.keys(allCards))
-        for (let key in collection) {
-          if (collection[key] && Number.isInteger(collection[key])) {
-            if (Object.keys(allCards).includes(key)) {
-              let thisCard = allCards[key].attributes
-              let thisCardsSet = cardSets[thisCard.set]
-              if (!Object.keys(playerCardCounts).includes(thisCard.set)) {
-                playerCardCounts[thisCard.set] = thisCardsSet.counts
-                playerCardCounts[thisCard.set].name = thisCard.set
-                playerCardCounts[thisCard.set].mythicOwned = 0
-                playerCardCounts[thisCard.set].rareOwned = 0
-                playerCardCounts[thisCard.set].uncommonOwned = 0
-                playerCardCounts[thisCard.set].commonOwned = 0
-              }
-
-              if (thisCard.rarity == "Mythic Rare") {
-                playerCardCounts[thisCard.set].mythicOwned += collection[key];
-              } else if (thisCard.rarity == "Rare") {
-                playerCardCounts[thisCard.set].rareOwned += collection[key];
-              } else if (thisCard.rarity == "Uncommon") {
-                playerCardCounts[thisCard.set].uncommonOwned += collection[key];
-              } else if (thisCard.rarity == "Common") {
-                playerCardCounts[thisCard.set].commonOwned += collection[key];
-              }
-              playerCardCounts[thisCard.set]
-            }
-            total += collection[key]
-            unique += 1
-          }
-        }
-
-        postMessage({
-          lastCollectionCount: `${unique} unique cards, ${total} total cards`,
-          playerCardCounts: Object.values(playerCardCounts),
-        })
-      }
-    }
-    // for(;;) {}; // use this to loop forever, and test that hung worker doesn't make window hang
-  })
-
-  collectionWorker.onmessage = event => {
-    if (event.data.lastCollectionCount) {
-      settingsData.lastCollectionCount = event.data.lastCollectionCount
-      settingsData.lastCollectionSetProgress = event.data.playerCardCounts
-    } else if (event.data.ready) {
-      collectionWorker.postMessage({lastCollection: settingsData.lastCollection})
-    }
-  }
-  collectionWorker.postMessage({allCards: mtga.allCards})
 
   let themePath = settingsData.runFromSource ? "themes" : path.join("..", "themes");
   fs.readdir(themePath, (err, files) => {
@@ -853,34 +603,30 @@ document.addEventListener("DOMContentLoaded", function(event) {
       })
     })
   })
-  $("#exportCollectionMTGGButton").click((e) => {
-    console.log("exporting mtgg to desktop")
-    let allPromises = []
-    for (let cardKey in settingsData.lastCollection) {
-      allPromises.push(mtga.allCards.findCard(cardKey))
-    }
 
-    Promise.all(allPromises).then(allCards => {
-      let mtggExportPath = path.join(desktopPath, 'mtga_collection_mtggoldfish.csv')
-      let csvContents = "Name,Edition,Qty,Foil\n"
-      for (let card of allCards) {
-        if (card) {
-          let mtgaID = card.get("mtgaID")
-          let prettyName = card.get("prettyName")
-          let set = card.get("set")
-          if (set == "DAR") set = "DOM"  // sigh, c'mon arena devs
-          let count = settingsData.lastCollection[mtgaID]
-          csvContents +=`"${prettyName}",${set},${count},No\n`
-        }
-      }
-      fs.writeFile(mtggExportPath, csvContents, (err) => {
-        if (err) {
-          alert(`error saving export: ${err}`)
-        } else {
-          alert(`Saved to ${mtggExportPath} !`)
-        }
+  $("#send-feedback").click((e) => {
+    console.log("sending feedback")
+    let data = {
+      feedbackType: $("#feedback-type-select").val(),
+      feedbackText: $("#feedback-text").val(),
+      contactInfo: $("#contact-info").val(),
+    }
+    if (!data.feedbackText) {
+      alert("Whoops, you forgot to fill out... er, the most important part of the form!")
+    } else {
+      $("#send-feedback").prop("disabled", true)
+      $("#send-feedback").html("Working on it...")
+      passThrough("public-api/feedback", data).then(r => {
+        console.log("success")
+        $("#send-feedback").html("Sent! Thanks!")
+      }).catch(e => {
+        console.log("error sending feedback :(")
+        console.log(e)
+        alert("Something went wrong! Try again?")
+        $("#send-feedback").prop("disabled", false)
+        $("#send-feedback").html("Try again?")
       })
-    })
+    }
   })
 
   document.getElementById("hide-delay").value = "" + settingsData.hideDelay;
@@ -912,25 +658,6 @@ document.addEventListener("DOMContentLoaded", function(event) {
     $(".slidevalue-vault").html(value)
   }
 
-  document.getElementById("recent-cards-quantity-slider").value = "" + settingsData.recentCardsQuantityToShow;
-  let initialValueRecentCardsQuantityToShow = settingsData.recentCardsQuantityToShow;
-  if(initialValueRecentCardsQuantityToShow == 100) {
-    initialValueRecentCardsQuantityToShow = "∞"
-  }
-  $(".slidevalue-recent-cards").html(initialValueRecentCardsQuantityToShow)
-  document.getElementById("recent-cards-quantity-slider").onchange = function() {
-    let value = parseInt(this.value)
-    settingsData.recentCardsQuantityToShow = value;
-    ipcRenderer.send('settingsChanged', {key: "recentCardsQuantityToShow", value: value})
-  }
-  document.getElementById("recent-cards-quantity-slider").oninput = function() {
-    let value = this.value
-    settingsData.recentCardsQuantityToShow = value;
-    if(value == 100) {
-      value = "∞"
-    }
-    $(".slidevalue-recent-cards").html(value);
-  }
 })
 
 // https://hackernoon.com/functional-javascript-resolving-promises-sequentially-7aac18c4431e
@@ -975,3 +702,31 @@ let getTrackerToken = () => {
 }
 
 // ipcRenderer.send('settingsChanged', {cool: true})
+var uploadDelay = 0;
+function passThrough(endpoint, passData, errors) {
+  if (!errors) {
+    errors = []
+  }
+  return new Promise((resolve, reject) => {
+    setTimeout(() => {
+      console.log(`posting ${endpoint} request...`)
+      request.post({
+        url: `${API_URL}/${endpoint}`,
+        json: true,
+        body: passData,
+        headers: {'User-Agent': 'MTGATracker-App'}
+      }, (err, res, data) => {
+        console.log(`finished posting ${endpoint} request...`)
+        if (err || res.statusCode != 200) {
+          errors.push({on: `post_${endpoint}`, error: err || res})
+          reject({errors: errors})
+        } else {
+          console.log(`${endpoint} uploaded! huzzah!`)
+          resolve({
+            success: true
+          })
+        }
+      })
+    }, 100 * uploadDelay)
+  })
+}
